Finish integrating guzzle
