+++ src/Provider/AbstractProvider.php
@@ -2,13 +2,23 @@
 
 namespace App\Provider;
 
+use App\Builder;
+use App\Facades\Log;
+use App\Facades\Settings;
+use App\Model\Deployment;
 use App\Model\Project;
+use Closure;
+use Exception;
 use GuzzleHttp\ClientInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
 use ReflectionClass;
+use Ronanchilvers\Foundation\Config;
 use Ronanchilvers\Utility\Str;
 use RuntimeException;
+use Symfony\Component\Process\Exception\ProcessFailedException;
+use Symfony\Component\Process\Process;
+use Symfony\Component\Yaml\Yaml;
 
 /**
  * Base provider class
@@ -162,25 +172,121 @@ abstract class AbstractProvider
     }
 
     /**
-     * Get a curl handle
-     *
-     * @param string $url
-     * @return resource
-     * @author Ronan Chilvers <ronan@d3r.com>
+     * @see \App\Provider\ProviderInterface::scanConfiguration()
      */
-    protected function getCurlHandle($url)
+    public function scanConfiguration(Project $project, Deployment $deployment, Closure $closure = null)
     {
-        if (!$curl = curl_init($url)) {
-            throw new RuntimeException('Unable to initialise CURL Github API request');
+        $repository = $this->encodeRepository($project->repository);
+        $params = [
+            'repository' => $repository,
+            'sha'        => $deployment->sha,
+        ];
+        $url = Str::moustaches(
+            $this->configUrl,
+            $params
+        );
+        $data = $this->getJSON($url);
+        $yaml = base64_decode($data['content']);
+        try {
+            $yaml = Yaml::parse($yaml);
+            $closure(
+                'info',
+                implode("\n", [
+                    'YAML deployment configuration read successfully',
+                    "JSON: " . json_encode($data, JSON_PRETTY_PRINT)
+                ])
+            );
+        } catch (Exception $ex) {
+            $closure(
+                'error',
+                implode("\n", [
+                    'Unable to parse YAML deployment configuration',
+                    "Exception: " . $ex->getMessage(),
+                ])
+            );
+            Log::error('Unable to parse YAML deployment configuration', [
+                'project'   => $project->toArray(),
+                'exception' => $ex,
+            ]);
+            return;
         }
-        curl_setopt_array($curl, [
-            CURLOPT_USERAGENT      => 'ronanchilvers/deploy - curl ' . curl_version()['version'],
-            CURLOPT_FOLLOWLOCATION => false,
-            CURLOPT_RETURNTRANSFER => true,
-            CURLOPT_TIMEOUT        => 5,
-        ]);
 
-        return $curl;
+        return new Config($yaml);
+    }
+
+    /**
+     * @see \App\Provider\ProviderInterface::download()
+     */
+    public function download(Project $project, Deployment $deployment, $directory, Closure $closure = null)
+    {
+        $repository = $this->encodeRepository($project->repository);
+        $params = [
+            'repository' => $repository,
+            'sha'        => $deployment->sha,
+        ];
+        $url = Str::moustaches(
+            $this->downloadUrl,
+            $params
+        );
+        $filename = tempnam('/tmp', Str::join('-', 'deploy', $project->id, $params['sha']));
+        if (!$handle = fopen($filename, "w")) {
+            $closure(
+                'error',
+                implode("\n", [
+                    "Unable to open temporary download file: {$filename}"
+                ])
+            );
+            throw new RuntimeException('Unable to open temporary file');
+        }
+        $response = $this->get(
+            $url,
+            [
+                'sink' => $handle,
+            ]
+        );
+
+        // Make sure the deployment download directory exists
+        if (!is_dir($directory)) {
+            $mode = Settings::get('build.chmod.default_folder', Builder::MODE_DEFAULT);
+            $closure('info', "Creating deployment directory: {$directory}");
+            if (!mkdir($directory, $mode, true)) {
+                $closure('error', "Failed to create deployment directory: {$directory}");
+                throw new RuntimeException(
+                    'Unable to create build directory at ' . $directory
+                );
+            }
+        }
+
+        // Decompress the archive into the download directory
+        $tar     = Settings::get('binary.tar', '/bin/tar');
+        $command = "{$tar} --strip-components=1 -xzf {$filename} -C {$directory}";
+        $closure('info', "Unpacking codebase tarball: {$command}");
+        $process = new Process(explode(' ', $command));
+        $process->run();
+        if (!$process->isSuccessful()) {
+            $closure(
+                'error',
+                implode("\n", [
+                    "Unpack failed: {$command}",
+                    $process->getErrorOutput(),
+                ])
+            );
+            throw new ProcessFailedException($process);
+        }
+
+        // Remove the downloaded archive
+        if (!unlink($filename)) {
+            $closure(
+                'error',
+                implode("\n", [
+                    'Unable to remove tarball after unpacking',
+                    $process->getOutput(),
+                ])
+            );
+            throw new RuntimeException('Unable to remove local code archive');
+        }
+
+        return true;
     }
 
     /**
@@ -228,5 +334,16 @@ abstract class AbstractProvider
     protected function client(): ClientInterface
     {
         return $this->client;
+    }
+
+    /**
+     * Encode a repository name
+     *
+     * @return string
+     * @author Ronan Chilvers <ronan@d3r.com>
+     */
+    protected function encodeRepository($repository)
+    {
+        return $repository;
     }
 }
